+++ src/anemoi/transform/transform.py
@@ -8,8 +8,6 @@
 # nor does it submit to any jurisdiction.
 
 
-import inspect
-import textwrap
 from abc import ABC
 from abc import ABCMeta
 from abc import abstractmethod
@@ -18,6 +16,8 @@ from typing import Callable
 from typing import TypeVar
 
 import earthkit.data as ekd
+
+from ..documentation import documentation_for_filter
 
 T = TypeVar("T", bound="Transform")
 
@@ -150,96 +150,19 @@ class Transform(ABC, metaclass=_TransformMetaClass):
 
     @classmethod
     def documentation(cls, filter_name) -> str:
+        """Returns the documentation for the transform.
 
-        from io import StringIO
+        Parameters
+        ----------
+        filter_name : str
+            The name of the filter.
 
-        from ruamel.yaml import YAML
-        from ruamel.yaml.comments import CommentedMap
-
-        yaml = YAML()
-        yaml.indent(sequence=4, offset=2)
-
-        def _lines(s: str) -> str:
-            lines = s.splitlines()
-            if len(lines) <= 1:
-                return lines
-            indent0 = len(lines[0]) - len(lines[0].lstrip())
-            indent1 = len(lines[1]) - len(lines[1].lstrip())
-            if indent0 == indent1:
-                return lines
-            return [lines[0]] + [line[indent1:] for line in lines[1:]]
-
-        result = _lines(cls.__doc__ or "")
-
-        examples = []
-        examples.append("")
-        examples.append("Examples")
-        examples.append("--------")
-        examples.append("")
-        examples.append(
-            """
-To use this filter in a dataset recipe, include it as show below, adjusting parameters as needed.
-See the `anemoi-datasets documentation <https://anemoi.readthedocs.io/>`_ for more details.
-"""
-        )
-
-        def _(annotation: Any) -> str:  # simple string representation of type annotations
-            if hasattr(annotation, "__name__"):
-                return annotation.__name__
-            return str(annotation).replace("typing.", "")
-
-        sig = inspect.signature(cls.__init__)
-        params = CommentedMap({})
-        for name, param in sig.parameters.items():
-            if name == "self":
-                continue
-            if param.default is inspect.Parameter.empty:
-                params[name] = "..."
-                params.yaml_add_eol_comment(f"{_(param.annotation)} (REQUIRED)", name)
-            else:
-                params[name] = param.default
-                params.yaml_add_eol_comment(f"{_(param.annotation)}", name)
-
-        dataset_example = CommentedMap(
-            {
-                "input": CommentedMap(
-                    {
-                        "pipe": [
-                            s := CommentedMap(
-                                {
-                                    "source": {
-                                        "param1": "value1",
-                                        "param2": "value2",
-                                        "param3": "...",
-                                    }
-                                }
-                            ),
-                            CommentedMap({filter_name: params}),
-                        ]
-                    }
-                )
-            }
-        )
-
-        s.yaml_add_eol_comment("Replace `source` with actual data source, e.g., 'mars', 'file', etc.", "source")
-
-        buf = StringIO()
-        yaml.dump(dataset_example, buf)
-        dataset_example = buf.getvalue()
-
-        # assert False, dataset_example
-
-        dataset_example = textwrap.indent(dataset_example, "  ")
-
-        examples.append(".. code-block:: yaml")
-        examples.append("")
-        examples.append(dataset_example)
-
-        examples.append("")
-
-        result.extend(examples)
-
-        return "\n".join(result)
+        Returns
+        -------
+        str
+            The documentation for the transform.
+        """
+        return documentation_for_filter(cls, filter_name)
 
 
 class ReversedTransform(Transform):
